Updated line chart with axis limit and label arguments and line a color argument
renderLineChart has been updated to return a list.  The renderValue binding now expects a list with at least one element named 'data'.  This is where the data.frame is expected.  Optional list elements are ylim, xlim, ylab, xlab, and cols.

Colors format is interpreted and converted to hex; accepted color formats are hex, rgb (as a vector), and R color names.  Colors are assigned to the corresponding line when the data.frame is converted to a JSON compatible list.

Axis limits (xlim and ylim) are expected as a vector of length 2, [i.e. c(min, max)].  Axis labels (xlab and ylab) are expected as type character.  Axis parameters are simply passed through to the renderValue binding.