Support for Horizontal Bar Charts
Additional support for horizontal bar charts.  Data should be sent to renderHorizontalBarChart as a list with at least one element named 'data'.  The data element is expected as a data frame with a column for group labels and a column of data for each group.  The labels column can be anywhere if it is named 'label', otherwise it must be the first column.

Additional optional settings:
- seriesNames: Character vector of names to be used in place of colnames of data
- cols: vector of colors to be used in place of nvd3 default colors.  Accepts rgb vectors, r color names, or hex codes.
- mar: Vector specifying margins of chart area (top, right, bottom, left)
- xlab: X axis label.  NOTE: nvd3 specifies the x axis as the y axis for horizontal bar charts.  As such the xlab value is used to set the axisLabel attribute of chart.yAxis.  This is not a typo.
- xlim: X axis limits.  See note on xlab.
- showTooltip: Boolean specifying whether the tooltips should be enabled.
- showValues: Boolean specifying whether the values should be added to the chart.  Values are not shown on 'stacked' layout.
- showControls: Boolean specifying whether controls for stacked/grouped should be shown.